Split parameters utils file out to have a validations utils file
