+++ components/utils/parameters.ts
@@ -1,4 +1,4 @@
-import { type Parameter, type ParameterOption, type ParameterSet, TypeOfParameter, type ValueData } from "~/types/parameterTypes";
+import type { Parameter, ParameterOption, ParameterSet, ValueData } from "~/types/parameterTypes";
 import type { Option } from "vue3-select-component";
 
 export interface ParameterSelectOption extends Option<string> {
@@ -55,32 +55,4 @@ export const getRangeForDependentParam = (
   if (dependentParam.updateNumericFrom && typeof dependedOnParamInputVal !== "undefined") {
     return dependentParam.updateNumericFrom?.values[dependedOnParamInputVal.toString()];
   }
-};
-
-// Given a numeric parameter where the ValueData is dependent on another parameter's value,
-// and the current values for each parameter (which could be data in a form, or a current scenario),
-// check whether the numeric value is out of the range defined by the metadata.
-export const numericValueIsOutOfRange = (
-  value: string | undefined,
-  parameter: Parameter | undefined,
-  parameterValueSet: ParameterSet | undefined,
-) => {
-  if (!value || parameter?.parameterType !== TypeOfParameter.Numeric) {
-    return false;
-  }
-
-  const int = Number.parseInt(value);
-  const range = getRangeForDependentParam(parameter, parameterValueSet);
-
-  return range && (int < range.min || int > range.max);
-};
-
-export const numericValueInvalid = (
-  value: string,
-  parameter: Parameter,
-) => {
-  if (parameter.parameterType === TypeOfParameter.Numeric) {
-    return Number(value) < 0 || Number.isNaN(Number(value));
-  };
-  return false;
 };
